refactor: test data in separate file
