+++ test/plots/parallel-coordinates.ts
@@ -1,285 +1,5 @@
 import { minMaxScaling, parallelCoordinates } from "../../src";
-
-function getRandomInt(max: number) {
-  return Math.floor(Math.random() * max);
-}
-
-interface DC {
-  type: string;
-  status: string;
-  power: number | string;
-  waterUsage: number | string;
-  surface: number;
-}
-
-const dcData: DC[] = [
-  {
-    type: "colocation",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "closed",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "closed",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "closed",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "colocation",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "private",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "private",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "private",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "private",
-    status: "open",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "private",
-    status: "closed",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  },
-  {
-    type: "hyperscaler",
-    status: "project",
-    power: getRandomInt(10),
-    waterUsage: getRandomInt(15),
-    surface: getRandomInt(500)
-  }
-];
+import { DC, dcData } from "../data/data";
 
 export function renderParallelCoordinates() {
   const plotId = "parallel-coordinates-plot";
